requirements.txt
# Web framework
fastapi==0.115.12
uvicorn[standard]==0.34.3

# Database
sqlalchemy==2.0.41
alembic==1.15.2
asyncpg==0.30.0        # async PostgreSQL driver

# Settings
pydantic==2.11.3
pydantic-settings==2.9.1

# Auth
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
passlib[bcrypt]==1.7.4

# Rate limiting
slowapi==0.1.9

# HTTP client (Google OAuth)
httpx==0.28.1

# Email templates
jinja2==3.1.6

# File uploads
python-multipart==0.0.20

# Testing
pytest==8.3.5
pytest-asyncio==0.26.0
aiosqlite==0.20.0      # async SQLite for tests